+++ CHANGELOG.rst
@@ -21,6 +21,7 @@ Added
     - ``get_query()`` helper for ``SpecificQuery`` initialization (#1003)
 - **Landingzones**
     - Zone UUID clipboard copying link (#970)
+    - Add ``inactivezones`` management command (#1010)
 - **Ontologyaccess**
     - Add site app for ontology storage and access (#937, #947)
     - Add ``importobo`` and ``importomim`` management commands (#937, #980)